requirement.txt
torch==2.8.0
transformers==4.35.2
datasets==2.14.6
pillow>=10.3.0
numpy
pyyaml==6.0.1